+++ resources/views/people/helpers/export.blade.php
@@ -1,26 +1,20 @@
-<table>
-    <thead>
-        <tr>
-            @foreach($fields as $field)
-                <th>
-                    @lang($field['label_key'])
-                </th>
-            @endforeach
-        </tr>
-    </thead>
-    <tbody>
-        @foreach($helpers as $helper)
-            <tr>
-                @foreach($fields as $field)
-                    <td>
-                        @if(gettype($field['value']) == 'string')
-                            {{ $field['prefix'] ?? '' }}{{ $helper->{$field['value']} }}
-                        @else
-                            {{ $field['prefix'] ?? '' }}{{ $field['value']($helper) }}
-                        @endif
-                    </td>
-                @endforeach
-            </tr>
-        @endforeach
-    </tbody>
-</table>
+@extends('layouts.app')
+
+@section('title', __('people.exportieren_helper_data'))
+
+@section('content')
+
+    {!! Form::open(['route' => 'people.helpers.doExport']) !!}
+        <div class="mb-3">
+            {{ Form::bsRadioList('format', $formats, $format, __('app.file_format')) }}
+        </div>
+        <div class="mb-3">
+            {{ Form::bsRadioList('scope', $scopes, $scope, __('people.scope')) }}
+        </div>
+        <p>
+            {{ Form::bsSubmitButton(__('app.export'), 'download') }}
+        </p>
+    {!! Form::close() !!}
+    
+@endsection
+