+++ Changelog.md
@@ -8,6 +8,7 @@ Helpers:
 * Gender grouping
 * Null values in certain groupings
 * WhatsApp link adds persons name
+* Export formats and scope
 
 ## 1.21.3